refactor: rename dev package as feature-dev
Rename the dev feature package directory and workspace package name to @workspace/feature-dev.

Update feature-shell imports, package dependencies, and the Bun lockfile to use the new package name.
