+++ packages/feature-dev/src/ui/dev-db-ui-record-details.tsx
