+++ packages/feature-dev/src/ui/dev-db-ui-table-record-field-value.tsx
