+++ packages/feature-shell/package.json
@@ -7,8 +7,8 @@
     "@workspace/background": "workspace:*",
     "@workspace/db": "workspace:*",
     "@workspace/db-react": "workspace:*",
-    "@workspace/dev": "workspace:*",
     "@workspace/env": "workspace:*",
+    "@workspace/feature-dev": "workspace:*",
     "@workspace/feature-explorer": "workspace:*",
     "@workspace/feature-onboarding": "workspace:*",
     "@workspace/feature-portfolio": "workspace:*",
