+++ packages/feature-shell/src/ui/shell-ui-menu.tsx
@@ -5,7 +5,7 @@ import { useNetworkLive } from '@workspace/db-react/use-network-live'
 import { useSetting } from '@workspace/db-react/use-setting'
 import { useWalletActive } from '@workspace/db-react/use-wallet-active'
 import { useWalletLive } from '@workspace/db-react/use-wallet-live'
-import { getDevOptions } from '@workspace/dev/dev-features'
+import { getDevOptions } from '@workspace/feature-dev/dev-features'
 import { Menubar } from '@workspace/ui/components/menubar'
 import { ShellUiMenuDevelopment } from './shell-ui-menu-development.tsx'
 import { ShellUiMenuNetwork } from './shell-ui-menu-network.tsx'
